add relation depth degree 1 or 2 to simplify relation graphs
New parameter relationdegreedepth, choose the relation depth degree 1 or 2 to simplify relation graphs for database with lot of tables and relations

+++ src/main/java/org/schemaspy/output/html/mustache/diagrams/MustacheTableDiagramFactory.java
@@ -42,28 +42,38 @@ public class MustacheTableDiagramFactory {
     private final MustacheDiagramFactory mustacheDiagramFactory;
     private final File outputDir;
     private final File tableDir;
+    private int relationDegreeDepth = 2;
 
-    public MustacheTableDiagramFactory(DotFormatter dotProducer, MustacheDiagramFactory mustacheDiagramFactory, File outputDir) {
+    public MustacheTableDiagramFactory(DotFormatter dotProducer, MustacheDiagramFactory mustacheDiagramFactory, File outputDir, int relationDegreeDepth) {
         this.dotProducer = dotProducer;
         this.mustacheDiagramFactory = mustacheDiagramFactory;
         this.outputDir = outputDir;
         this.tableDir = outputDir.toPath().resolve("diagrams").resolve("tables").toFile();
         tableDir.mkdirs();
+        this.relationDegreeDepth = relationDegreeDepth;
     }
 
     public List<MustacheTableDiagram> generateTableDiagrams(Table table, WriteStats stats) throws IOException {
         List<MustacheTableDiagram> diagrams = new ArrayList<>();
+        
         File oneDegreeDotFile = new File(tableDir, table.getName() + ".1degree.dot");
-        File twoDegreesDotFile = new File(tableDir, table.getName() + ".2degrees.dot");
         File oneImpliedDotFile = new File(tableDir, table.getName() + ".implied1degrees.dot");
-        File twoImpliedDotFile = new File(tableDir, table.getName() + ".implied2degrees.dot");
+        
+        File twoDegreesDotFile = null;
+        File twoImpliedDotFile = null;
+        
+        if (relationDegreeDepth == 2) {
+        	twoDegreesDotFile = new File(tableDir, table.getName() + ".2degrees.dot");
+            twoImpliedDotFile = new File(tableDir, table.getName() + ".implied2degrees.dot");
+            Files.deleteIfExists(twoDegreesDotFile.toPath());
+            Files.deleteIfExists(twoImpliedDotFile.toPath());
+        }
 
         // delete before we start because we'll use the existence of these files to determine
         // if they should be turned into pngs & presented
         Files.deleteIfExists(oneDegreeDotFile.toPath());
-        Files.deleteIfExists(twoDegreesDotFile.toPath());
         Files.deleteIfExists(oneImpliedDotFile.toPath());
-        Files.deleteIfExists(twoImpliedDotFile.toPath());
+        
 
 
         if (table.getMaxChildren() + table.getMaxParents() > 0) {
@@ -77,38 +87,42 @@ public class MustacheTableDiagramFactory {
             oneDiagram.setActive(true);
             diagrams.add(oneDiagram);
 
-            WriteStats twoStats = new WriteStats(stats);
-            try (PrintWriter dotOut = Writers.newPrintWriter(twoDegreesDotFile)) {
-                impliedConstraints = dotProducer.writeRealRelationships(table, true, twoStats, dotOut, outputDir);
+            WriteStats oneImplied = new WriteStats(stats);
+            try (PrintWriter dotOut = Writers.newPrintWriter(oneImpliedDotFile)) {
+                dotProducer.writeAllRelationships(table, false, oneImplied, dotOut, outputDir);
             }
+            MustacheTableDiagram oneImpliedDiagram = mustacheDiagramFactory.generateTableDiagram("One implied", oneImpliedDotFile, table.getName() + ".implied1degrees");
+            oneImpliedDiagram.setIsImplied(true);
+            diagrams.add(oneImpliedDiagram);
+ 
 
-            if (sameWritten(oneStats, twoStats)) {
-                Files.deleteIfExists(twoDegreesDotFile.toPath()); // no different than before, so don't show it
-            } else {
-                diagrams.add(mustacheDiagramFactory.generateTableDiagram("Two degrees", twoDegreesDotFile, table.getName() + ".2degrees"));
-            }
-
-            if (!impliedConstraints.isEmpty()) {
-                WriteStats oneImplied = new WriteStats(stats);
-                try (PrintWriter dotOut = Writers.newPrintWriter(oneImpliedDotFile)) {
-                    dotProducer.writeAllRelationships(table, false, oneImplied, dotOut, outputDir);
-                }
-                MustacheTableDiagram oneImpliedDiagram = mustacheDiagramFactory.generateTableDiagram("One implied", oneImpliedDotFile, table.getName() + ".implied1degrees");
-                oneImpliedDiagram.setIsImplied(true);
-                diagrams.add(oneImpliedDiagram);
-                WriteStats twoImplied = new WriteStats(stats);
-                try (PrintWriter dotOut = Writers.newPrintWriter(twoImpliedDotFile)) {
-                    dotProducer.writeAllRelationships(table, true, twoImplied, dotOut, outputDir);
-                }
-                if (sameWritten(oneImplied, twoImplied)) {
-                    Files.deleteIfExists(twoImpliedDotFile.toPath());
-                } else {
-                    MustacheTableDiagram twoImpliedDiagram = mustacheDiagramFactory.generateTableDiagram("Two implied", twoImpliedDotFile, table.getName() + ".implied2degrees");
-                    twoImpliedDiagram.setIsImplied(true);
-                    diagrams.add(twoImpliedDiagram);
-                }
-                return diagrams;
-            }
+	        if (relationDegreeDepth == 2) {
+	            WriteStats twoStats = new WriteStats(stats);
+	            try (PrintWriter dotOut = Writers.newPrintWriter(twoDegreesDotFile)) {
+	                impliedConstraints = dotProducer.writeRealRelationships(table, true, twoStats, dotOut, outputDir);
+	            }
+	
+	            if (sameWritten(oneStats, twoStats)) {
+	                Files.deleteIfExists(twoDegreesDotFile.toPath()); // no different than before, so don't show it
+	            } else {
+	                diagrams.add(mustacheDiagramFactory.generateTableDiagram("Two degrees", twoDegreesDotFile, table.getName() + ".2degrees"));
+	            }
+         
+	            if(!impliedConstraints.isEmpty()) {
+	                WriteStats twoImplied = new WriteStats(stats);
+	                try (PrintWriter dotOut = Writers.newPrintWriter(twoImpliedDotFile)) {
+	                    dotProducer.writeAllRelationships(table, true, twoImplied, dotOut, outputDir);
+	                }
+	                if (sameWritten(oneImplied, twoImplied)) {
+	                    Files.deleteIfExists(twoImpliedDotFile.toPath());
+	                } else {
+	                    MustacheTableDiagram twoImpliedDiagram = mustacheDiagramFactory.generateTableDiagram("Two implied", twoImpliedDotFile, table.getName() + ".implied2degrees");
+	                    twoImpliedDiagram.setIsImplied(true);
+	                    diagrams.add(twoImpliedDiagram);
+	                }
+	                return diagrams;
+	            }
+	        }
         }
 
         return diagrams;
